Revert "Arnej/use common hamming distance"
