+++ vespalib/CMakeLists.txt
@@ -139,7 +139,6 @@ vespa_define_module(
     src/tests/util/file_area_freelist
     src/tests/util/generationhandler
     src/tests/util/generationhandler_stress
-    src/tests/util/hamming
     src/tests/util/md5
     src/tests/util/mmap_file_allocator
     src/tests/util/mmap_file_allocator_factory
