--- vespalib/src/tests/util/hamming/CMakeLists.txt
@@ -1,9 +1,0 @@
-# Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-vespa_add_executable(vespalib_hamming_test_app TEST
-    SOURCES
-    hamming_test.cpp
-    DEPENDS
-    vespalib
-    GTest::GTest
-)
-vespa_add_test(NAME vespalib_hamming_test_app COMMAND vespalib_hamming_test_app)
