+++ vespalib/src/vespa/vespalib/util/CMakeLists.txt
@@ -11,7 +11,6 @@ vespa_add_library(vespalib_vespalib_util OBJECT
     barrier.cpp
     benchmark_timer.cpp
     bfloat16.cpp
-    binary_hamming_distance.cpp
     blockingthreadstackexecutor.cpp
     box.cpp
     child_process.cpp
